Added the possibility to flag a Face as frequently updating (dynamic).
Added a Variable, which allows to specify,  if an object should expect frequent changes.
Based upon that variable, an object will decide, weather to push changes directly to the Graphics card, or wait till the frame is rendered, to allow additional changes to be processed, before any data is sent.